Code request
Hi I am a student studying AI as it relates to economics and finance at Northwestern and came across your work on Medium. This approach to using GANs in a forecasting setting is fascinating to me. I would love to play around with your code, if possible, to better learn you implementation method. Thanks!