Please help with GAN and reinforcement learning
Hi Boris,
My and my friend are preparing applications for graduate school and we are so interested in the Artificial intelligence area and we decide to make a project related to that. Then we found your paper.
Your work about GAN and reinforcement learning on hyperparameters was quite attractive! I am really interested in learning how a genius like you think this kind of problem and try to find related patterns to apply to it.
Could you please share with me the reinforcement learning idea and perhaps also the GAN code and deep reinforcement learning code that you used to analyze the stock data?
I would sincerely appreciate your help!

Thank you very much,
Frank Kang
kynkangyining@gmail.com